requirements.txt
autopep8==1.6.0
pika==1.2.0
pkg_resources==0.0.0
pycodestyle==2.8.0
python-dotenv==0.19.2
toml==0.10.2
